chore: Separates the functional and snapshot tests in dirs
