+++ src/__tests__/functional-tests/form-field-controls-integration.test.tsx
@@ -3,11 +3,11 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 
-import FormField, { FormFieldProps } from '../../lib/components/form-field';
-import { FormFieldValidationControlProps } from '../../lib/components/internal/context/form-field-context';
-import createWrapper, { ElementWrapper } from '../../lib/components/test-utils/dom';
-import { getRequiredPropsForComponent } from './required-props-for-components';
-import { requireComponent } from './utils';
+import FormField, { FormFieldProps } from '../../../lib/components/form-field';
+import { FormFieldValidationControlProps } from '../../../lib/components/internal/context/form-field-context';
+import createWrapper, { ElementWrapper } from '../../../lib/components/test-utils/dom';
+import { getRequiredPropsForComponent } from '../required-props-for-components';
+import { requireComponent } from '../utils';
 
 const formFieldControlComponents = [
   {
@@ -77,7 +77,7 @@ formFieldControlComponents.forEach(({ componentName, findNativeElement }) => {
     const requiredProps = getRequiredPropsForComponent(componentName) ?? {};
     const renderResult = render(
       <FormField {...formFieldProps}>
-        <Component {...requiredProps} {...controlComponentProps} value="" onChange={() => {}} />
+        <Component {...requiredProps} {...controlComponentProps} value="" onChange={() => { }} />
       </FormField>
     );
     const formFieldWrapper = createWrapper(renderResult.container).findFormField()!;
