+++ src/__tests__/functional-tests/naming-convention.test.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 import { pascalCase } from 'change-case';
 
-import { getAllComponents, requireComponent } from './utils';
+import { getAllComponents, requireComponent } from '../utils';
 
 describe('component displayName matches import directory name', () => {
   getAllComponents().forEach(componentName => {
