+++ src/__tests__/functional-tests/outer-form-submit.test.tsx
@@ -3,8 +3,8 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 
-import { getRequiredPropsForComponent } from './required-props-for-components';
-import { getAllComponents, requireComponent } from './utils';
+import { getRequiredPropsForComponent } from '../required-props-for-components';
+import { getAllComponents, requireComponent } from '../utils';
 
 const skippedComponents = ['button'];
