+++ src/__tests__/functional-tests/public-exports.test.ts
@@ -5,7 +5,7 @@ import path from 'path';
 
 // There are no typings for package.json
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const packageJson = require('../../lib/components/package.json');
+const packageJson = require('../../../lib/components/package.json');
 
 test('all exports declarations resolve to a file', () => {
   for (const exportPath of Object.values<string>(packageJson.exports)) {
@@ -13,6 +13,8 @@ test('all exports declarations resolve to a file', () => {
     if (exportPath === './internal/context/') {
       continue;
     }
-    expect(() => require.resolve(exportPath, { paths: [path.join(__dirname, '../../lib/components')] })).not.toThrow();
+    expect(() =>
+      require.resolve(exportPath, { paths: [path.join(__dirname, '../../../lib/components')] })
+    ).not.toThrow();
   }
 });
