+++ src/__tests__/functional-tests/use-base-component.test.tsx
@@ -6,9 +6,9 @@ import { pascalCase } from 'change-case';
 
 import { COMPONENT_METADATA_KEY } from '@cloudscape-design/component-toolkit/internal';
 
-import { PACKAGE_VERSION } from '../../lib/components/internal/environment';
-import { getRequiredPropsForComponent } from './required-props-for-components';
-import { getAllComponents, requireComponent, supportsDOMProperties } from './utils';
+import { PACKAGE_VERSION } from '../../../lib/components/internal/environment';
+import { getRequiredPropsForComponent } from '../required-props-for-components';
+import { getAllComponents, requireComponent, supportsDOMProperties } from '../utils';
 
 describe('useBaseComponent hook is used in all allowlisted public components', () => {
   const componentRoot = document.createElement('div');
