+++ src/__tests__/snapshot-tests/design-tokens.test.ts
@@ -1,6 +1,6 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import { requireDesignTokensFile } from './utils';
+import { requireDesignTokensFile } from '../utils';
 
 const themeNames = ['classic', 'visual-refresh'];
