+++ src/__tests__/snapshot-tests/documenter.test.ts
@@ -1,6 +1,6 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import { getAllComponents, requireComponentDefinition } from './utils';
+import { getAllComponents, requireComponentDefinition } from '../utils';
 
 describe('Documenter', () => {
   test.each<string>(getAllComponents())(`definition for %s matches the snapshot`, (componentName: string) => {
